Remove key delay tracking
It's going to be replaced with 'held' and 'pressed' methods

the 'held' method will work the same as the current 'pressed'
method does; if the user is pressing the key, it returns true.

the new 'pressed' method will return only one true for each key
press, resetting when the key is released.